+++ apis/autoscalingplans/v1beta2/zz_scalingplan_types.go
@@ -234,10 +234,10 @@ type ScalingInstructionInitParameters struct {
 	DisableDynamicScaling *bool `json:"disableDynamicScaling,omitempty" tf:"disable_dynamic_scaling,omitempty"`
 
 	// Maximum capacity of the resource. The exception to this upper limit is if you specify a non-default setting for predictive_scaling_max_capacity_behavior.
-	MaxCapacity *float64 `json:"maxCapacity,omitempty" tf:"max_capacity,omitempty"`
+	MaxCapacity *int64 `json:"maxCapacity,omitempty" tf:"max_capacity,omitempty"`
 
 	// Minimum capacity of the resource.
-	MinCapacity *float64 `json:"minCapacity,omitempty" tf:"min_capacity,omitempty"`
+	MinCapacity *int64 `json:"minCapacity,omitempty" tf:"min_capacity,omitempty"`
 
 	// Predefined load metric to use for predictive scaling. You must specify either predefined_load_metric_specification or customized_load_metric_specification when configuring predictive scaling.
 	// More details can be found in the AWS Auto Scaling API Reference.
@@ -248,7 +248,7 @@ type ScalingInstructionInitParameters struct {
 	PredictiveScalingMaxCapacityBehavior *string `json:"predictiveScalingMaxCapacityBehavior,omitempty" tf:"predictive_scaling_max_capacity_behavior,omitempty"`
 
 	// Size of the capacity buffer to use when the forecast capacity is close to or exceeds the maximum capacity.
-	PredictiveScalingMaxCapacityBuffer *float64 `json:"predictiveScalingMaxCapacityBuffer,omitempty" tf:"predictive_scaling_max_capacity_buffer,omitempty"`
+	PredictiveScalingMaxCapacityBuffer *int64 `json:"predictiveScalingMaxCapacityBuffer,omitempty" tf:"predictive_scaling_max_capacity_buffer,omitempty"`
 
 	// Predictive scaling mode. Valid values: ForecastAndScale, ForecastOnly.
 	PredictiveScalingMode *string `json:"predictiveScalingMode,omitempty" tf:"predictive_scaling_mode,omitempty"`
@@ -263,7 +263,7 @@ type ScalingInstructionInitParameters struct {
 	ScalingPolicyUpdateBehavior *string `json:"scalingPolicyUpdateBehavior,omitempty" tf:"scaling_policy_update_behavior,omitempty"`
 
 	// Amount of time, in seconds, to buffer the run time of scheduled scaling actions when scaling out.
-	ScheduledActionBufferTime *float64 `json:"scheduledActionBufferTime,omitempty" tf:"scheduled_action_buffer_time,omitempty"`
+	ScheduledActionBufferTime *int64 `json:"scheduledActionBufferTime,omitempty" tf:"scheduled_action_buffer_time,omitempty"`
 
 	// Namespace of the AWS service. Valid values: autoscaling, dynamodb, ecs, ec2, rds.
 	ServiceNamespace *string `json:"serviceNamespace,omitempty" tf:"service_namespace,omitempty"`
@@ -283,10 +283,10 @@ type ScalingInstructionObservation struct {
 	DisableDynamicScaling *bool `json:"disableDynamicScaling,omitempty" tf:"disable_dynamic_scaling,omitempty"`
 
 	// Maximum capacity of the resource. The exception to this upper limit is if you specify a non-default setting for predictive_scaling_max_capacity_behavior.
-	MaxCapacity *float64 `json:"maxCapacity,omitempty" tf:"max_capacity,omitempty"`
+	MaxCapacity *int64 `json:"maxCapacity,omitempty" tf:"max_capacity,omitempty"`
 
 	// Minimum capacity of the resource.
-	MinCapacity *float64 `json:"minCapacity,omitempty" tf:"min_capacity,omitempty"`
+	MinCapacity *int64 `json:"minCapacity,omitempty" tf:"min_capacity,omitempty"`
 
 	// Predefined load metric to use for predictive scaling. You must specify either predefined_load_metric_specification or customized_load_metric_specification when configuring predictive scaling.
 	// More details can be found in the AWS Auto Scaling API Reference.
@@ -297,7 +297,7 @@ type ScalingInstructionObservation struct {
 	PredictiveScalingMaxCapacityBehavior *string `json:"predictiveScalingMaxCapacityBehavior,omitempty" tf:"predictive_scaling_max_capacity_behavior,omitempty"`
 
 	// Size of the capacity buffer to use when the forecast capacity is close to or exceeds the maximum capacity.
-	PredictiveScalingMaxCapacityBuffer *float64 `json:"predictiveScalingMaxCapacityBuffer,omitempty" tf:"predictive_scaling_max_capacity_buffer,omitempty"`
+	PredictiveScalingMaxCapacityBuffer *int64 `json:"predictiveScalingMaxCapacityBuffer,omitempty" tf:"predictive_scaling_max_capacity_buffer,omitempty"`
 
 	// Predictive scaling mode. Valid values: ForecastAndScale, ForecastOnly.
 	PredictiveScalingMode *string `json:"predictiveScalingMode,omitempty" tf:"predictive_scaling_mode,omitempty"`
@@ -312,7 +312,7 @@ type ScalingInstructionObservation struct {
 	ScalingPolicyUpdateBehavior *string `json:"scalingPolicyUpdateBehavior,omitempty" tf:"scaling_policy_update_behavior,omitempty"`
 
 	// Amount of time, in seconds, to buffer the run time of scheduled scaling actions when scaling out.
-	ScheduledActionBufferTime *float64 `json:"scheduledActionBufferTime,omitempty" tf:"scheduled_action_buffer_time,omitempty"`
+	ScheduledActionBufferTime *int64 `json:"scheduledActionBufferTime,omitempty" tf:"scheduled_action_buffer_time,omitempty"`
 
 	// Namespace of the AWS service. Valid values: autoscaling, dynamodb, ecs, ec2, rds.
 	ServiceNamespace *string `json:"serviceNamespace,omitempty" tf:"service_namespace,omitempty"`
@@ -335,11 +335,11 @@ type ScalingInstructionParameters struct {
 
 	// Maximum capacity of the resource. The exception to this upper limit is if you specify a non-default setting for predictive_scaling_max_capacity_behavior.
 	// +kubebuilder:validation:Optional
-	MaxCapacity *float64 `json:"maxCapacity" tf:"max_capacity,omitempty"`
+	MaxCapacity *int64 `json:"maxCapacity" tf:"max_capacity,omitempty"`
 
 	// Minimum capacity of the resource.
 	// +kubebuilder:validation:Optional
-	MinCapacity *float64 `json:"minCapacity" tf:"min_capacity,omitempty"`
+	MinCapacity *int64 `json:"minCapacity" tf:"min_capacity,omitempty"`
 
 	// Predefined load metric to use for predictive scaling. You must specify either predefined_load_metric_specification or customized_load_metric_specification when configuring predictive scaling.
 	// More details can be found in the AWS Auto Scaling API Reference.
@@ -353,7 +353,7 @@ type ScalingInstructionParameters struct {
 
 	// Size of the capacity buffer to use when the forecast capacity is close to or exceeds the maximum capacity.
 	// +kubebuilder:validation:Optional
-	PredictiveScalingMaxCapacityBuffer *float64 `json:"predictiveScalingMaxCapacityBuffer,omitempty" tf:"predictive_scaling_max_capacity_buffer,omitempty"`
+	PredictiveScalingMaxCapacityBuffer *int64 `json:"predictiveScalingMaxCapacityBuffer,omitempty" tf:"predictive_scaling_max_capacity_buffer,omitempty"`
 
 	// Predictive scaling mode. Valid values: ForecastAndScale, ForecastOnly.
 	// +kubebuilder:validation:Optional
@@ -373,7 +373,7 @@ type ScalingInstructionParameters struct {
 
 	// Amount of time, in seconds, to buffer the run time of scheduled scaling actions when scaling out.
 	// +kubebuilder:validation:Optional
-	ScheduledActionBufferTime *float64 `json:"scheduledActionBufferTime,omitempty" tf:"scheduled_action_buffer_time,omitempty"`
+	ScheduledActionBufferTime *int64 `json:"scheduledActionBufferTime,omitempty" tf:"scheduled_action_buffer_time,omitempty"`
 
 	// Namespace of the AWS service. Valid values: autoscaling, dynamodb, ecs, ec2, rds.
 	// +kubebuilder:validation:Optional
@@ -412,7 +412,7 @@ type ScalingPlanObservation struct {
 	ScalingInstruction []ScalingInstructionObservation `json:"scalingInstruction,omitempty" tf:"scaling_instruction,omitempty"`
 
 	// The version number of the scaling plan. This value is always 1.
-	ScalingPlanVersion *float64 `json:"scalingPlanVersion,omitempty" tf:"scaling_plan_version,omitempty"`
+	ScalingPlanVersion *int64 `json:"scalingPlanVersion,omitempty" tf:"scaling_plan_version,omitempty"`
 }
 
 type ScalingPlanParameters struct {
@@ -478,7 +478,7 @@ type TargetTrackingConfigurationInitParameters struct {
 
 	// Estimated time, in seconds, until a newly launched instance can contribute to the CloudWatch metrics.
 	// This value is used only if the resource is an Auto Scaling group.
-	EstimatedInstanceWarmup *float64 `json:"estimatedInstanceWarmup,omitempty" tf:"estimated_instance_warmup,omitempty"`
+	EstimatedInstanceWarmup *int64 `json:"estimatedInstanceWarmup,omitempty" tf:"estimated_instance_warmup,omitempty"`
 
 	// Predefined metric. You can specify either predefined_scaling_metric_specification or customized_scaling_metric_specification.
 	// More details can be found in the AWS Auto Scaling API Reference.
@@ -486,11 +486,11 @@ type TargetTrackingConfigurationInitParameters struct {
 
 	// Amount of time, in seconds, after a scale in activity completes before another scale in activity can start.
 	// This value is not used if the scalable resource is an Auto Scaling group.
-	ScaleInCooldown *float64 `json:"scaleInCooldown,omitempty" tf:"scale_in_cooldown,omitempty"`
+	ScaleInCooldown *int64 `json:"scaleInCooldown,omitempty" tf:"scale_in_cooldown,omitempty"`
 
 	// Amount of time, in seconds, after a scale-out activity completes before another scale-out activity can start.
 	// This value is not used if the scalable resource is an Auto Scaling group.
-	ScaleOutCooldown *float64 `json:"scaleOutCooldown,omitempty" tf:"scale_out_cooldown,omitempty"`
+	ScaleOutCooldown *int64 `json:"scaleOutCooldown,omitempty" tf:"scale_out_cooldown,omitempty"`
 
 	// Target value for the metric.
 	TargetValue *float64 `json:"targetValue,omitempty" tf:"target_value,omitempty"`
@@ -507,7 +507,7 @@ type TargetTrackingConfigurationObservation struct {
 
 	// Estimated time, in seconds, until a newly launched instance can contribute to the CloudWatch metrics.
 	// This value is used only if the resource is an Auto Scaling group.
-	EstimatedInstanceWarmup *float64 `json:"estimatedInstanceWarmup,omitempty" tf:"estimated_instance_warmup,omitempty"`
+	EstimatedInstanceWarmup *int64 `json:"estimatedInstanceWarmup,omitempty" tf:"estimated_instance_warmup,omitempty"`
 
 	// Predefined metric. You can specify either predefined_scaling_metric_specification or customized_scaling_metric_specification.
 	// More details can be found in the AWS Auto Scaling API Reference.
@@ -515,11 +515,11 @@ type TargetTrackingConfigurationObservation struct {
 
 	// Amount of time, in seconds, after a scale in activity completes before another scale in activity can start.
 	// This value is not used if the scalable resource is an Auto Scaling group.
-	ScaleInCooldown *float64 `json:"scaleInCooldown,omitempty" tf:"scale_in_cooldown,omitempty"`
+	ScaleInCooldown *int64 `json:"scaleInCooldown,omitempty" tf:"scale_in_cooldown,omitempty"`
 
 	// Amount of time, in seconds, after a scale-out activity completes before another scale-out activity can start.
 	// This value is not used if the scalable resource is an Auto Scaling group.
-	ScaleOutCooldown *float64 `json:"scaleOutCooldown,omitempty" tf:"scale_out_cooldown,omitempty"`
+	ScaleOutCooldown *int64 `json:"scaleOutCooldown,omitempty" tf:"scale_out_cooldown,omitempty"`
 
 	// Target value for the metric.
 	TargetValue *float64 `json:"targetValue,omitempty" tf:"target_value,omitempty"`
@@ -539,7 +539,7 @@ type TargetTrackingConfigurationParameters struct {
 	// Estimated time, in seconds, until a newly launched instance can contribute to the CloudWatch metrics.
 	// This value is used only if the resource is an Auto Scaling group.
 	// +kubebuilder:validation:Optional
-	EstimatedInstanceWarmup *float64 `json:"estimatedInstanceWarmup,omitempty" tf:"estimated_instance_warmup,omitempty"`
+	EstimatedInstanceWarmup *int64 `json:"estimatedInstanceWarmup,omitempty" tf:"estimated_instance_warmup,omitempty"`
 
 	// Predefined metric. You can specify either predefined_scaling_metric_specification or customized_scaling_metric_specification.
 	// More details can be found in the AWS Auto Scaling API Reference.
@@ -549,12 +549,12 @@ type TargetTrackingConfigurationParameters struct {
 	// Amount of time, in seconds, after a scale in activity completes before another scale in activity can start.
 	// This value is not used if the scalable resource is an Auto Scaling group.
 	// +kubebuilder:validation:Optional
-	ScaleInCooldown *float64 `json:"scaleInCooldown,omitempty" tf:"scale_in_cooldown,omitempty"`
+	ScaleInCooldown *int64 `json:"scaleInCooldown,omitempty" tf:"scale_in_cooldown,omitempty"`
 
 	// Amount of time, in seconds, after a scale-out activity completes before another scale-out activity can start.
 	// This value is not used if the scalable resource is an Auto Scaling group.
 	// +kubebuilder:validation:Optional
-	ScaleOutCooldown *float64 `json:"scaleOutCooldown,omitempty" tf:"scale_out_cooldown,omitempty"`
+	ScaleOutCooldown *int64 `json:"scaleOutCooldown,omitempty" tf:"scale_out_cooldown,omitempty"`
 
 	// Target value for the metric.
 	// +kubebuilder:validation:Optional
