+++ apis/cloudwatchevents/v1beta1/zz_archive_types.go
@@ -35,7 +35,7 @@ type ArchiveInitParameters struct {
 	EventSourceArnSelector *v1.Selector `json:"eventSourceArnSelector,omitempty" tf:"-"`
 
 	// The maximum number of days to retain events in the new event archive. By default, it archives indefinitely.
-	RetentionDays *float64 `json:"retentionDays,omitempty" tf:"retention_days,omitempty"`
+	RetentionDays *int64 `json:"retentionDays,omitempty" tf:"retention_days,omitempty"`
 }
 
 type ArchiveObservation struct {
@@ -55,7 +55,7 @@ type ArchiveObservation struct {
 	ID *string `json:"id,omitempty" tf:"id,omitempty"`
 
 	// The maximum number of days to retain events in the new event archive. By default, it archives indefinitely.
-	RetentionDays *float64 `json:"retentionDays,omitempty" tf:"retention_days,omitempty"`
+	RetentionDays *int64 `json:"retentionDays,omitempty" tf:"retention_days,omitempty"`
 }
 
 type ArchiveParameters struct {
@@ -89,7 +89,7 @@ type ArchiveParameters struct {
 
 	// The maximum number of days to retain events in the new event archive. By default, it archives indefinitely.
 	// +kubebuilder:validation:Optional
-	RetentionDays *float64 `json:"retentionDays,omitempty" tf:"retention_days,omitempty"`
+	RetentionDays *int64 `json:"retentionDays,omitempty" tf:"retention_days,omitempty"`
 }
 
 // ArchiveSpec defines the desired state of Archive
